- Admin Navbar => in Material UI - updated Header of container app. - clean up of this branch so that only code related to navbar would be in this branch
