+++ Dashboard/src/App.js
@@ -1,14 +1,12 @@
 import React from "react"
 import { BrowserRouter } from "react-router-dom"
-import Navigation from "./routes-nav/Navigation"
 import Routes from "./routes-nav/Routes"
 
 function App(history) {
   return (
     <BrowserRouter>
       <div className="App">
-        <Navigation />
-        <Routes />
+        <Routes/>
       </div>
     </BrowserRouter>
   )
