+++ Dashboard/src/routes-nav/Routes.js
@@ -7,7 +7,6 @@ import EditEntity from "../EntityDetails/EditEntity"
 import Loan from "../Loan"
 import Regulatory from "../Regulatory"
 import MyRequest from "../MyRequest"
-import Confirmation from "../Confirmation"
 
 /** Site-wide routes.
  * Visiting a non-existant route redirects to the homepage.
@@ -17,7 +16,7 @@ function Routes() {
   return (
     <div className="pt-5">
       <Switch>
-        <Route exact path="/">
+        <Route path="/home">
           <Homepage />
         </Route>
 
@@ -43,10 +42,6 @@ function Routes() {
 
         <Route exact path="/myrequest">
           <MyRequest />
-        </Route>
-
-        <Route exact path="/confirmation">
-          <Confirmation />
         </Route>
 
         <Redirect to="/" />
